+++ tests/Feature/ClientUtilsTest.php
@@ -44,101 +44,4 @@ class ClientUtilsTest extends TestCase
 
         $client->disconnect();
     }
-
-    public function test_loop_event_handlers_are_called_for_each_loop_iteration(): void
-    {
-        $client = new MqttClient($this->mqttBrokerHost, $this->mqttBrokerPort, 'test-loop-event-handler');
-
-        $loopCount           = 0;
-        $previousElapsedTime = 0;
-        $client->registerLoopEventHandler(function (MqttClient $client, float $elapsedTime) use (&$loopCount, &$previousElapsedTime) {
-            $this->assertGreaterThanOrEqual($previousElapsedTime, $elapsedTime);
-
-            $previousElapsedTime = $elapsedTime;
-
-            $loopCount++;
-
-            if ($loopCount >= 3) {
-                $client->interrupt();
-                return;
-            }
-        });
-
-        $client->connect();
-
-        $client->loop();
-
-        $this->assertSame(3, $loopCount);
-
-        $client->disconnect();
-    }
-
-    public function test_loop_event_handler_can_be_unregistered_and_will_not_be_called_anymore(): void
-    {
-        $client = new MqttClient($this->mqttBrokerHost, $this->mqttBrokerPort, 'test-loop-event-handler');
-
-        $loopCount = 0;
-        $handler = function (MqttClient $client) use (&$loopCount) {
-            $loopCount++;
-
-            if ($loopCount >= 1) {
-                $client->interrupt();
-                return;
-            }
-        };
-
-        $client->registerLoopEventHandler($handler);
-
-        $client->connect();
-
-        $client->loop();
-
-        $this->assertSame(1, $loopCount);
-
-        $client->unregisterLoopEventHandler($handler);
-
-        $client->loop(true, true);
-
-        $this->assertSame(1, $loopCount);
-
-        $client->disconnect();
-    }
-
-    public function test_all_loop_event_handlers_can_be_unregistered_and_will_not_be_called_anymore(): void
-    {
-        $client = new MqttClient($this->mqttBrokerHost, $this->mqttBrokerPort, 'test-loop-event-handler');
-
-        $loopCount1 = 0;
-        $loopCount2 = 0;
-        $handler1 = function (MqttClient $client) use (&$loopCount1) {
-            $loopCount1++;
-
-            if ($loopCount1 >= 1) {
-                $client->interrupt();
-                return;
-            }
-        };
-        $handler2 = function () use (&$loopCount2) {
-            $loopCount2++;
-        };
-
-        $client->registerLoopEventHandler($handler1);
-        $client->registerLoopEventHandler($handler2);
-
-        $client->connect();
-
-        $client->loop();
-
-        $this->assertSame(1, $loopCount1);
-        $this->assertSame(1, $loopCount2);
-
-        $client->unregisterLoopEventHandler();
-
-        $client->loop(true, true);
-
-        $this->assertSame(1, $loopCount1);
-        $this->assertSame(1, $loopCount2);
-
-        $client->disconnect();
-    }
 }
